requirements.txt
django-cors-headers==3.6.0
graphene-django
django-admin-interface